Edited the name of utils and added tests for it
